Moving of RDConfig to riffdog.config plus refactor of command entry
